New offline logic

+++ app/src/main/java/com/infomaniak/drive/data/sync/UploadAdapter.kt
@@ -211,7 +211,7 @@ class UploadAdapter @JvmOverloads constructor(
                     }
                     startUploadFile(uploadFile, cacheFile.length(), syncResult)
                     UploadFile.deleteIfExists(uri)
-                    cacheFile.delete()
+                    if (!uploadFile.isSyncOffline()) cacheFile.delete()
                 } else {
                     SyncUtils.checkDocumentProviderPermissions(context, uri)
                     val fileSize = try {
